Scope experiment viewer to the visualization package
Removes this change's framework-side edits so it touches only isaaclab_arena/visualization/,
leaving the Experiment manifest to a follow-up change.

- Drop experiment_manifest.py and reconstruct_experiment_manifest.py. The report groups Runs by
  factorizing Run names instead, which reproduces the manifest's grouping on the 38-task
  experiment exactly (39 tasks x 2 policies, 23409 episodes).
- Drop run_status.py and episode_video_files.py, reverting arena_run.py, experiment_runner.py and
  camera_observation_video_recorder.py to main. These only existed so the report could be imported
  without the simulation stack, which no in-tree caller needs.
- Move episode_results_files.py into the visualization package, its only consumer.
