+++ isaaclab_arena/video/camera_observation_video_recorder.py
@@ -26,20 +26,50 @@ from __future__ import annotations
 import gymnasium as gym
 import numpy as np
 import os
+import re
 import torch
 from dataclasses import dataclass
 
 from moviepy.video.io.ffmpeg_writer import FFMPEG_VideoWriter
 
-from isaaclab_arena.video.episode_video_files import (
-    ParsedEpisodeVideoName,
-    format_episode_video_filename,
-    parse_episode_video_filename,
+CAMERA_OBS_GROUP_KEY = "camera_obs"
+
+# Regular expression to parse the filename of an episode video.
+_EPISODE_VIDEO_FILENAME_PATTERN = re.compile(
+    r"^(?P<prefix>.+?)(?:-rebuild(?P<rebuild>\d+))?-env(?P<env>\d+)-(?P<camera>.+)-episode-(?P<episode>\d+)\.mp4$"
 )
 
-__all__ = ["ParsedEpisodeVideoName", "format_episode_video_filename", "parse_episode_video_filename"]
 
-CAMERA_OBS_GROUP_KEY = "camera_obs"
+@dataclass
+class ParsedEpisodeVideoName:
+    """The fields recovered from a recorder mp4 filename by ``parse_episode_video_filename``."""
+
+    prefix: str
+    env_index: int
+    camera_name: str
+    episode_index: int
+    rebuild_index: int | None
+    """The rebuild this video belongs to, or ``None`` when the prefix carried no ``-rebuild`` segment."""
+
+
+def format_episode_video_filename(name_prefix: str, env_index: int, camera_name: str, episode_index: int) -> str:
+    """Build the mp4 filename for one (env, camera, episode). Inverse of ``parse_episode_video_filename``."""
+    return f"{name_prefix}-env{env_index}-{_sanitize_cam_key(camera_name)}-episode-{episode_index}.mp4"
+
+
+def parse_episode_video_filename(filename: str) -> ParsedEpisodeVideoName | None:
+    """Parse a recorder mp4 filename, or return ``None`` if it does not match the recorder's format."""
+    match = _EPISODE_VIDEO_FILENAME_PATTERN.match(filename)
+    if match is None:
+        return None
+    rebuild = match.group("rebuild")
+    return ParsedEpisodeVideoName(
+        prefix=match.group("prefix"),
+        env_index=int(match.group("env")),
+        camera_name=match.group("camera"),
+        episode_index=int(match.group("episode")),
+        rebuild_index=int(rebuild) if rebuild is not None else None,
+    )
 
 
 def _to_uint8(frame: torch.Tensor | np.ndarray) -> np.ndarray:
@@ -52,6 +82,11 @@ def _to_uint8(frame: torch.Tensor | np.ndarray) -> np.ndarray:
         scale = 255.0 if float(frame.max()) <= 1.0 else 1.0
         return np.clip(frame * scale, 0, 255).astype(np.uint8)
     return frame.astype(np.uint8)
+
+
+def _sanitize_cam_key(camera_name: str) -> str:
+    """Strip path separators so a camera name can't escape video_folder."""
+    return camera_name.replace("/", "_").replace(os.sep, "_")
 
 
 @dataclass
